Final Stream-lit/requirements.txt
streamlit==0.64.0
numpy==1.19.1
pandas==0.25.0
joblib==0.16.0
matplotlib==3.3.0
scikit-learn==0.23.1
scipy==1.5.2
boruta==0.3
seaborn==0.10.1
plotly==4.9.0
pystan==2.19.1.1
prophet==0.1.1.post1
feature-engine==0.5.14
pytorch==1.0.2